memdb: Fix a bug in Put() API
Put() is expected to be a no-op if the item already exist. It is currently
broken. If you lookup in the skiplist for a item with insCmp, it will locate
an exactly equal item with same bornSn. If the item with exact bornSn does not exist,
the predecessor in the path buffer can be with exact same key, but with different bornSn.
If the predecessor is with the same key, but with old bornSn, we need
to check if it is dead. If the deadSn=0, that means an entry already exists
for the given key and insert should be skipped.