memdb: Do not try to read items from empty gclist
The BarrierSessionFlush may invoke the barrier session destructor
with an empty gclist. If the destructor sends the empty list to freeWorker,
it tries to interpret that as head node of the list. It leads to crash.
The destructor should check if the list is empty before passing it to the
freeWorker.