plasma: Implement index comparator and careful icmp based page split
We need to make sure that each page has unique set of keys. When items
are modified with MVCC sn, it is possible that pages are split with a
middle key which is differed by sn. Hence, page compact operations
will break as delete operation and insert operation on same key may be
split across two separate pages.

This change adds an index compare method to prevent this case from happening.
Index compare (icmp) is used to locate a page. The item compare (xcmp) defines
the ordering of items within a page for items with same keys, but different sn.

Change-Id: I12af8e993e1f468faabf264e2db7393f2ab9d258