plasma: Introduce lssPageUpdate block type to optimize cleaner and recovery
An LSS new page is represented by lssPageData or lssPageReloc. Subsequent
delta updates are enclosed in lssPageUpdate block. When cleaner runs,
it should only look for lssPageData or lssPageReloc. If cleaner has seen
an newpage block, that means it must have already relocated the page to
the end of the log already. Hence, it is not needed to process lssPageUpdate.

Similarly, whenever lssPageData or lssPageReloc is seen, recovery should
create a new index or replace the page data.

Change-Id: I018a32f1b8960670fa498de107f5c382f4e3d240