plasma: Implement safe log segment garbage collection
When a thread obtains a page, the page may contain a
swapout delta which points to a location in the LSS.
The log cleaner independently cleans log segments.
If a there is an inflight thread which has visibility
to a log segment, lss cleaner should not remove the file
until the thread goes away. We cannot use any lock based
reference counting as it will significantly affect the
performance. This commit implements a lockfree scheme similar
to the EPOCH based garbage collection.

Change-Id: Icb8516991626d88133ca2afede72fa5912e6f3e7