plasma: Fix indirection into cleaned page segment during lss relocate
If we add a relocation delta to a page with swapout delta, a reader
may try to read from lss through the swapout delta offset. That lss
location might be cleaned away by the lss cleaner already.

Hence, we have to force a page compaction to avoid this situation.
It is possible to optimize it so that we compact only if there is
a swapout delta in the page.

Change-Id: I561fae8b0148dc29f5cb4977e70de7ab71bf102f