MB-23082 plasma: Handle race conditions during page merge/remove
A page merge is a two step atomic operation.
Step 1: Mark page as removed
Step 2: Perform merge + delete index skiplist node

During step 2 is attempted, many things can happen concurrently
so that candidate parent page may not be pointing to the removed
node as a split on parent could be performed by a concurrent thread.
Similarly, the candidate parent itself could be marked for removal.
This can lead of cascaded removal of pages in the order of first page
in the chain. if tryPageRemoval is attempted as a best effort single
call, it may lead to situation causing livelock.

The following invarients are required for a successful merge:

1. An index lookup on parent page should ensure that its rightsibling is
the removed node.
2. If an index lookup returns a pid, we should verify that it is the same
page as the requested page removal. (A concurrent split could be adding a
page with same pid after a successful removal of the page).
3. A merge should not be performed is the candidate parent page is marked for
removal
4. The tryPageRemoval should return only after completion of page merge either
by the current routine or some other routine.

Change-Id: Ibc847a5a540d9a9491f387b7dbe8d86b8cf0d0a1