MB-22433 plasma: Fix a race condition in concurrent page delete + split insert
A page split consists of two steps:
1. Atomically updating the parent page with split delta
2. Inserting new PID to the index layer

Since these two operations are atomic, it is possible that insert into the
index layer can happen in a delayed manner. But, even though the new split page
is not yet visible in index layer, we still allow page operations to reach the
new split node through redirect mechanism in the split delta.

This transient invisibility of newly created page can lead to a race condition
during page delete. The page removal mechanism is co-operative. Hence, if a page
removal operation is attempted for PID, it it does not find the page in the index
layer, it is assumed that some other thread successfully deleted the page. But,
this assumption is not true due to transient invisibility of page split node.

A page which is newly created can be marked for deleted due to merge even before
a successful insertion into the index layer is complete. This may cause deletion
thread to skip deletion of the page causing more complex issues and livelock arising
from cascaded deletes of pages. Another possibility of race condition is that a newly
split page is immediately deleted even before it is inserted into the index layer.
Hence, the deleter thread thinks that somebody else removed it already as the split
index layer inser is not yet complete. At the same time another thread could cause
a split again for the merged page with exactly same PID low key. If the new split
succeeds before old split's index layer insert, it will crash when old split PID
is attempted to insert in the index layer with panic "duplicate index node"

The commit enforces that we can merge/delete a page only after it is visible in
the index layer.

Change-Id: I9bb83c4b11b47a141ff8ec8032e82eb98df2144f