MB-23572 plasma: Fix split page index key selection during page split
The exact split pivot key used for forming the new page items should be
used as the index key for the new page. Because of MVCC garbage collection,
it is possible that item-0 of the new page is not equal to the split pivot key.
Hence, it may lead to a condition having missing ranges in the index which
does not have any pages mapped.

Change-Id: I0d0cf114558744738e12af706719c7f41a1ffd5f