MB-23953 Bring swapout page back to memory before page merge
If there is a swapout page, a page relocate operation should swapin
the page to memory before adding relocate delta as the swapout offset
becomes invalid after relocation.

An iterator spawns a tree of iterators. The tree branching happens when
merge-delta is encountered. Each branch in the delta tree should be self
countained. ie., if there is a swapin delta, it should satisfiy a corresponding
swapout delta in the same branch. If there is a swapin delta for a swapout delta
in a branch, the page walker for the branch will not notice the swapin delta
for the swapout delta in the branch and hence it may cause an invalid read.
(log starts at offset <xxx>, trying to read <xxx>).

Eg:

page_delta
page_swapin
page_delta
page_delta
merge_delta -> [page_delta_link]
page_delta
page_swapout

Page walker for the subtree below merge_delta will not notice page_swapin
above the merge_delta.

To avoid this situation, a swapin should be forced before page merge

Change-Id: I3a7691ce856da0e67a38aafdb76f8fd88caf2064