MB-23951 plasma: Fix memory corruption of page deltas during page merge
The tryRemovePage() method is called with page to be merged as argument.
The argument is a closed page. Now, the parent page is obtained by looking
up the index layer. A merge delta is added to the parent page and index node
link to the parent page is CASed.

Before performing the page merge, we have to make sure that both parent and
closed page has all its components in memory. Hence, we add a swapin delta for
swapout pages. Now, if the CAS operation to update parent page fails, the
allocated swapin deltas are freed immediately and the the tryRemovePage() body
is repeated. But, the closed page instance is not refreshed. ie., closed page
will have additional swapin delta we just freed since the CAS failed. The next
retry of the tryRemovePage() operation may succeed and it will successfully add
the swapin (freed one) into the parent page as part of merge. Few moments later,
the SMR mechanism will actually reclaim the memory form the swapin delta and
corrupt the delta which is parent of the merged page. Any future access into
this merged page may result in a crash.

The fix is to make the closed page immutable by avoiding the need for adding
swapin delta during tryPageRemoval(). Instead, we swapin the page as part of
page close. Hence, during the retries, closed page will remain unmodified.

Change-Id: I32d6b2d0c915763f9ae9ce4f84fa23b5c7c7a401