MB-25100 plasma: Fix safe-memory-reclaimer stalling due to throttling
When writes are throttled, it holds an SMR token. Hence, SMR memory freeing
is stalled until it comes out of throttling loop and releases SMR token.
This leads to huge memory spikes as scan threads may allocate memory.

This change also enables memory throttling for scan operations.
writerSMRBufferSize is changed from 500 to 50 to reducing SMR buffering
of to-be-freed objects.

Change-Id: Id987552b04a22e453663feca75e5ae472c800663